+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Avoid `None` attribute error detected by conda-forge CI. 
+* Updated Artists regitery with context for latest `compas`
 
 ### Removed